Remove explicit SocketGroups
