+++ io/src/main/scala/fs2/io/net/Datagram.scala
@@ -23,5 +23,11 @@ package fs2
 package io
 package net
 
-/** Provides support for UDP networking. */
-package object udp
+import com.comcast.ip4s.{IpAddress, SocketAddress}
+
+/** A single datagram to send to the specified remote address or received from the specified address.
+  *
+  * @param remote   remote party to send/receive datagram to/from
+  * @param bytes    data to send/receive
+  */
+final case class Datagram(remote: SocketAddress[IpAddress], bytes: Chunk[Byte])
