+++ io/src/test/scala/fs2/io/net/tls/DTLSSocketSuite.scala
@@ -29,24 +29,23 @@ import scala.concurrent.duration._
 import cats.effect.{IO, Resource}
 import com.comcast.ip4s._
 
-import fs2.io.net.udp.{Packet, Socket}
-
 class DTLSSocketSuite extends TLSSuite {
+  val network = Network.create[IO]
+
   group("DTLSSocket") {
     test("echo") {
       val msg = Chunk.array("Hello, world!".getBytes)
 
-      def address(s: Socket[IO]) =
+      def address(s: DatagramSocket[IO]) =
         Resource
           .eval(s.localAddress)
           .map(a => SocketAddress(ip"127.0.0.1", a.port))
 
       val setup = for {
         tlsContext <- Resource.eval(testTlsContext)
-        socketGroup <- Network[IO].udpSocketGroup
-        serverSocket <- socketGroup.open()
+        serverSocket <- network.openDatagramSocket()
         serverAddress <- address(serverSocket)
-        clientSocket <- socketGroup.open()
+        clientSocket <- network.openDatagramSocket()
         clientAddress <- address(clientSocket)
         tlsServerSocket <- tlsContext.dtlsServer(serverSocket, clientAddress, logger = logger)
         tlsClientSocket <- tlsContext.dtlsClient(clientSocket, serverAddress, logger = logger)
@@ -60,7 +59,7 @@ class DTLSSocketSuite extends TLSSuite {
               .foreach(serverSocket.write(_))
           val echoClient = Stream.eval {
             IO.sleep(500.millis) >>
-              clientSocket.write(Packet(serverAddress, msg)) >>
+              clientSocket.write(Datagram(serverAddress, msg)) >>
               clientSocket.read
           }
