+++ io/src/test/scala/fs2/io/net/tls/TLSSuite.scala
@@ -28,11 +28,14 @@ import cats.effect.IO
 
 abstract class TLSSuite extends Fs2Suite {
   def testTlsContext: IO[TLSContext[IO]] =
-    Network[IO].tlsContext.fromKeyStoreResource(
-      "keystore.jks",
-      "password".toCharArray,
-      "password".toCharArray
-    )
+    Network
+      .create[IO]
+      .tlsContext
+      .fromKeyStoreResource(
+        "keystore.jks",
+        "password".toCharArray,
+        "password".toCharArray
+      )
 
   val logger = None // Some((msg: String) => IO(println(s"\u001b[33m${msg}\u001b[0m")))
 }
